✨ add hugo blog logic+content for all blog pages
* use layouts/tags to style both /tags and /tags/foo pages
* use layouts/_default/list.html for homepage, /posts and /links blog
  pages
* use layouts/_default/single.html for /posts/foo or links/foo
* use shortcodes/img.html to make integrating cover image or other
  additional images easier
* use partials/pagination.html to navigate multiple pages if too many
  entries
* use layouts/archive/list.html for content on /archive

Content:

I spent several days working on learning the Hugo Go template Language and
applying things the way I wanted to present the content of the blog. It
took me a lot of trial and error, but eventually I got it to work.

But the most important thing is that I want the home page to show
articles with title,datestamp,tags,cover image, summary and read more
link. I want them to be showing mixed entries from both /posts and
/links sorted by latest date.
In a similiar vain I wanted content on /posts and /links to work. But of
course only showing their own entries.

Then for the archive I wanted a list where each entry is sorted
by year together with the title and the link to it. For the tags I also
wanted to keep it simple and just show all tags. And for the tags/foo
page to show date - Title of all related pages.

Content management and folder structure:

I was thinking of using datestamps or isosec for
content/posts/timestamps/index.md. Or a system that manages content like
i.e. content/posts/title/index.md. Mainly because I didn't know any better
at the time.

The timestamps as folder names only make things more complicated and are
of no value, as the front matter already has a date variable with an ISO
8601 timestamp.

The title in the folder structure was more because I didn't know how to
handle URLS differently from the folder structure. This can be set in
the hugo.toml configuration file. But that was never my decision, it's
just the standard way Hugo handles it.

So IMO it makes most sense to keep it simple so
/content/posts/n/index.md where each new post is n+1.

Why not just /content/posts/title.md? Because the extra folder allows to
store the images at the same place.
* Images will be saved with n-uuid.jpg or if it's a cover image
  n-cover.jpg. I was thinking of maybe adding the width x height info too,
  but I am not sure as of yet.

There is also the option of saving all images in /static/foo/*.jpg. But
I found it easier to have all the files related to a blog post in the
same folder.